% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\docType{package}
\name{lwafomps-package}
\alias{lwafomps}
\alias{lwafomps-package}
\title{Multivariate Normal Distribution with Low WAFOM Point Set}
\description{
Compute the distribution function of the multivariate normal
distribution for arbitrary limits and correlation matrices.
}

