% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{mvnorm}
\alias{mvnorm}
\title{Multivariate Normal Distribution with Low WAFOM Point Set}
\usage{
mvnorm(dimension, lower, upper, mean, coval, abseps, releps,
  confidenceLevel = 0.95, timeLimit = Inf, parallelism = 4)
}
\arguments{
\item{dimension}{dimension or number of variables.}

\item{lower}{the vector of lower limits.}

\item{upper}{the vector of upper limits.}

\item{mean}{the mean vector.}

\item{coval}{the corelation matrix.}

\item{abseps}{expected absolute error tolerance.}

\item{releps}{expected relative error tolerance.}

\item{confidenceLevel}{confidence level.}

\item{timeLimit}{time limit (seconds).}

\item{parallelism}{degree of parallelism for fast calculation.}
}
\value{
integrated value of multivariate normal distribution.
}
\description{
Compute the distribution function of the multivariate normal
distribution for arbitrary limits and correlation matrices.
}

